Add command to move lift and arm together
