+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/Lift.java
@@ -11,6 +11,7 @@ import com.arcrobotics.ftclib.hardware.motors.MotorEx;
 import com.arcrobotics.ftclib.hardware.motors.MotorGroup;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import org.firstinspires.ftc.robotcore.external.Telemetry;
+import org.firstinspires.ftc.teamcode.util.LiftPosition;
 
 @Config
 public class Lift extends SubsystemBase {
@@ -102,13 +103,21 @@ public class Lift extends SubsystemBase {
     }
 
     /**
+     * Set the height of the lift.
+     * @param pos the position to set
+     */
+    public void setHeight(@NonNull LiftPosition pos) {
+        pidf.setPIDF(KP, KI, KD, KF);
+        pidf.setSetPoint(pos.ticks() + offset.ticks());
+    }
+
+    /**
      * Sets the target level of the lift.
      * @param level the new level to set
      */
     public void setLevel(@NonNull Level level) {
         this.level = level;
-        pidf.setPIDF(KP, KI, KD, KF);
-        pidf.setSetPoint(level.pos.ticks() + offset.ticks());
+        setHeight(level.pos);
     }
 
     /**
@@ -165,33 +174,4 @@ public class Lift extends SubsystemBase {
         }
     }
 }
-class LiftPosition {
-    private double mm;
-    private double ticks;
-    private static final double TICKS_PER_MM =  384.5 / 112; // Motor tics per revolution divided by circumference of pulley.
 
-    public LiftPosition(double mm){
-        setMm(mm);
-    }
-    public void setMm(double mm){
-        this.mm = mm;
-        this.ticks = mmToTicks(mm);
-    }
-    public void setTicks(double ticks){
-        this.ticks = ticks;
-        this.mm = ticksToMm(ticks);
-    }
-    public double mm() {
-        return mm;
-    }
-    public double ticks() {
-        return ticks;
-    }
-
-    public static double mmToTicks(double mm){
-        return mm * TICKS_PER_MM;
-    }
-    public static double ticksToMm(double ticks){
-        return ticks / TICKS_PER_MM;
-    }
-}
